runtime: narrow Clock.nowNs to u64 + fix last_refill_ns sentinel
Per team-lead design review: match the existing simulation-timeline
invariant (`SimIO.now_ns: u64`, `Simulator.clock_ns: u64`,
`PendingEntry.deadline_ns: u64`) rather than `std.time.nanoTimestamp()`'s
`i128`. The whole daemon was already u64-ns underneath; my first cut
introduced an i128 island that forced 128-bit conversion at every
`EventLoop → SimIO` boundary, widened the `Clock` struct, and
foreclosed `std.atomic.Value(u64)` for any future caller wanting a
shared ns timestamp.

u64 nanoseconds since the Unix epoch wraps in **year ~2554**
(`2^64 ns ≈ 584.5 years`) — far past any reasonable daemon lifetime.
Documented in the file header. A new `@sizeOf(Clock) <= 16` regression
test guards the size invariant; if someone widens the sim variant back
to i128 we want a test failure, not a silent perf regression.

Changes:

  * `src/runtime/clock.zig`:
      - `Clock.sim` field: `i128` → `u64`
      - `nowNs()` returns `u64` (production reads
        `std.time.nanoTimestamp()` then `@intCast` to u64; safe for
        current and future centuries)
      - `simAtNs(u64)`, `setNs(u64)` — narrowed
      - `simAtSecs(i64)` / `simAtMs(i64)` clamp negative input to 0
        (sim time is non-negative)
      - `advance{Secs,Ms,Ns}(i64)` saturate at 0 on backward jumps
        (centralised in `applyDelta` helper using saturating add/sub)
      - Year-2554 invariant + size test added

  * `src/io/rate_limiter.zig`:
      - All `now_ns: i128` parameters → `u64`
      - `last_refill_ns: i128` → `last_refill_ns: ?u64`. The previous
        `== 0` magic-value sentinel was fragile: a sim test anchored
        at `t = 0` would re-seed on every refill and never credit
        elapsed time. `?u64` makes "uninitialised" unambiguous. Two
        new tests guard the corner case: `refill clock-going-backward
        is a no-op` and `refill across the t=0 anchor seeds rather
        than crediting`.

  * `tests/clock_random_determinism_test.zig`:
      - `post_clock_ns: i128` → `u64` to track the API change.

  * STATUS.md and progress-report updated to record the type-width
    decision and what was learned (Pattern: match existing project
    invariants over stdlib mirror conventions when the choice is
    cheap).

`zig build test` clean across 3 representative seeds; no production
behaviour change beyond the previously-fragile sentinel-zero case
that was broken under specific sim-test conditions.

+++ tests/clock_random_determinism_test.zig
@@ -148,7 +148,7 @@ test "rate-limit + key-generation pipeline is deterministic under sim sources" {
     const ScenarioOut = struct {
         first_consume: u64,
         key: [8]u8,
-        post_clock_ns: i128,
+        post_clock_ns: u64,
     };
 
     const scenario = struct {
